requirements.txt
Pillow==9.2.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.12.2
streamlit_lottie==0.0.5
